Fix lỗi

+++ TelerikWebApp1/NhapSoLieuThue.aspx.cs
@@ -1,9 +1,12 @@
 ﻿using NPOI.SS.UserModel;
 using NPOI.XSSF.UserModel;
+using OfficeOpenXml;
+using OfficeOpenXml.Style;
 using System;
 using System.Collections.Generic;
 using System.Configuration;
 using System.Data;
+using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -23,7 +26,7 @@ namespace TelerikWebApp1
         StringBuilder st = new StringBuilder();
         protected void Page_Load(object sender, EventArgs e)
         {
-            if(IsPostBack)
+            if (IsPostBack)
             {
 
             }
@@ -131,7 +134,7 @@ namespace TelerikWebApp1
                     string Extension = Path.GetExtension(FileUpload1.PostedFile.FileName);
                     if ((Extension == ".xls") || (Extension == ".xlsx"))
                     {
-                        FilePath = Server.MapPath("~/file/"+ FileName);
+                        FilePath = Server.MapPath("~/file/" + FileName);
                         FileUpload1.SaveAs(FilePath);
                         var data = ReadFromExcelfile(FilePath);
                         Session["File"] = FilePath;
@@ -157,10 +160,97 @@ namespace TelerikWebApp1
                 ClientScript.RegisterClientScriptBlock(this.GetType(), "", st.ToString(), true);
             }
         }
+        private void BindingFormatForExcel(ExcelWorksheet worksheet, List<SoLieuTuKhoBac> listItems)
+        {
+            // Set default width cho tất cả column
+            worksheet.DefaultColWidth = 10;
+            // Tự động xuống hàng khi text quá dài
+            worksheet.Cells.Style.WrapText = true;
+            // Tạo header
+            worksheet.Cells[1, 1].Value = "Mã số thuế";
+            worksheet.Cells[1, 2].Value = "Kỳ thuế";
+            worksheet.Cells[1, 3].Value = "Số tiền nộp";
+            worksheet.Cells[1, 4].Value = "Tiểu mục";
+            worksheet.Cells[1, 5].Value = "Mục";
+            worksheet.Cells[1, 6].Value = "Ngày nộp";
+            worksheet.Cells[1, 7].Value = "Mã cán bộ quản lý";
+            // Lấy range vào tạo format cho range đó ở đây là từ A1 tới Q1
+            using (var range = worksheet.Cells["A1:H1"])
+            {
+                // Set PatternType
+                range.Style.Fill.PatternType = ExcelFillStyle.DarkGray;
+                // Set Màu cho Background
+                range.Style.Fill.BackgroundColor.SetColor(Color.White);
+                //
+                range.Style.Font.Color.SetColor(Color.Black);
+                // Canh giữa cho các text
+                range.Style.HorizontalAlignment = ExcelHorizontalAlignment.Center;
+                // Set Font cho text  trong Range hiện tại
+                range.Style.Font.SetFromFont(new Font("Arial", 10));
+                // Set Border
+                range.Style.Border.Bottom.Style = ExcelBorderStyle.Thick;
+                // Set màu ch Border
+                range.Style.Border.Bottom.Color.SetColor(Color.Blue);
 
+            }
+
+            // Đỗ dữ liệu từ list vào 
+            for (int i = 0; i < listItems.Count; i++)
+            {
+                var item = listItems[i];
+                worksheet.Cells[i + 2, 1].Value = item.masothue;
+                worksheet.Cells[i + 2, 2].Value = item.KyThue;
+                worksheet.Cells[i + 2, 3].Value = item.SoTienNop;
+                worksheet.Cells[i + 2, 4].Value = item.tieumuc;
+                worksheet.Cells[i + 2, 5].Value = item.muc;
+                worksheet.Cells[i + 2, 6].Value = item.NgayNop;
+                worksheet.Cells[i + 2, 7].Value = item.macbql;
+            }
+        }
+        private Stream CreateExcelFile(List<SoLieuTuKhoBac> list,Stream stream = null)
+        {
+        
+            using (var excelPackage = new ExcelPackage(stream ?? new MemoryStream()))
+            {
+                // Tạo author cho file Excel
+                excelPackage.Workbook.Properties.Author = "Hanker";
+                // Tạo title cho file Excel
+                excelPackage.Workbook.Properties.Title = "EPP test background";
+                // thêm tí comments vào làm màu 
+                excelPackage.Workbook.Properties.Comments = "This is my fucking generated Comments";
+                // Add Sheet vào file Excel
+                excelPackage.Workbook.Worksheets.Add("First Sheet");
+                // Lấy Sheet bạn vừa mới tạo ra để thao tác 
+                var workSheet = excelPackage.Workbook.Worksheets[1];
+                // Đổ data vào Excel file
+
+                BindingFormatForExcel(workSheet, list);
+                excelPackage.Save();
+                return excelPackage.Stream;
+            }
+        }
+
+        public void Export(List<SoLieuTuKhoBac>lst)
+        {
+            // Gọi lại hàm để tạo file excel
+            var stream = CreateExcelFile(lst,null);
+            // Tạo buffer memory strean để hứng file excel
+            var buffer = stream as MemoryStream;
+            // Đây là content Type dành cho file excel, còn rất nhiều content-type khác nhưng cái này mình thấy okay nhất
+            Response.ContentType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
+            // Dòng này rất quan trọng, vì chạy trên firefox hay IE thì dòng này sẽ hiện Save As dialog cho người dùng chọn thư mục để lưu
+            // File name của Excel này là ExcelDemo
+            Response.AddHeader("Content-Disposition", "attachment; filename=DanhSachLoi.xlsx");
+            // Lưu file excel của chúng ta như 1 mảng byte để trả về response
+            Response.BinaryWrite(buffer.ToArray());
+            // Send tất cả ouput bytes về phía clients
+            Response.Flush();
+            Response.End();
+        }
         protected void btnSave_Click(object sender, EventArgs e)
         {
             GridDataItem item;
+            List<SoLieuTuKhoBac> lstErr = new List<SoLieuTuKhoBac>();
             st.Clear();
             for (int i = 0; i < grid.Items.Count; i++)
             {
@@ -180,15 +270,81 @@ namespace TelerikWebApp1
                         DanhBa danhba = db.DanhBas.SingleOrDefault(x => x.masothue == masothue);
                         if (danhba == null)
                             continue;
-                        db.Insert_SoLieuNopThue(masothue, macbql, sotiennop, ngaynop, kythue, tieumuc, muc);
-                        if ((System.IO.File.Exists(txtFilepath.Text)))
+                        else
                         {
-                            System.IO.File.Delete(txtFilepath.Text);
+                            KhaiThue kt = db.KhaiThues.SingleOrDefault(x => x.masothue == masothue);
+                            if (kt != null)
+                            {
+                                var sobo = db.SoBoThues.SingleOrDefault(x => x.idKhaiThue == kt.idKhaiThue);
+                                //Kiểm tra nếu nộp thiếu tiền thuế thì trả về danh sách các hộ có số thu chưa đúng 
+                                //Do chỉ có 2 trường hợp nộp đủ và không nộp
+                                if(tieumuc == "1003" && muc =="1000")
+                                {
+                                    if(sotiennop < sobo.SoTienTNCN1Thang)
+                                    {
+                                        lstErr.Add(new SoLieuTuKhoBac
+                                        {
+                                            masothue = masothue,
+                                            macbql = macbql,
+                                            SoTienNop = int.Parse(sotiennop.ToString()),
+                                            NgayNop = DateTime.Parse(ngaynop),
+                                            KyThue = kythue,
+                                            tieumuc = tieumuc,
+                                            muc= muc
+                                        });
+                                        continue;
+                                    }
+                                }
+                                else if(tieumuc=="1701" && muc == "1700")
+                                {
+                                    if (sotiennop < sobo.SoTienGTGT1Thang)
+                                    {
+                                        lstErr.Add(new SoLieuTuKhoBac
+                                        {
+                                            masothue = masothue,
+                                            macbql = macbql,
+                                            SoTienNop = int.Parse(sotiennop.ToString()),
+                                            NgayNop = DateTime.Parse(ngaynop),
+                                            KyThue = kythue,
+                                            tieumuc = tieumuc,
+                                            muc = muc
+                                        });
+                                        continue;
+                                    }
+                                }
+                                else
+                                {
+                                    thuemonbai monbai = db.thuemonbais.SingleOrDefault(x=>x.idKhaiThue == kt.idKhaiThue);
+                                    if(monbai !=null  && monbai.trangthai == false)
+                                    {
+                                        mucluc_MonBai mucluc_MB = db.mucluc_MonBais.SingleOrDefault(x => x.Bac == monbai.Bac);
+                                        if (sotiennop < mucluc_MB.MucThue)
+                                        {
+                                            lstErr.Add(new SoLieuTuKhoBac
+                                            {
+                                                masothue = masothue,
+                                                macbql = macbql,
+                                                SoTienNop = int.Parse(sotiennop.ToString()),
+                                                NgayNop = DateTime.Parse(ngaynop),
+                                                KyThue = kythue,
+                                                tieumuc = tieumuc,
+                                                muc = muc
+                                            });
+                                            continue;
+                                        }
+                                    }
+                                }
+                                db.Insert_SoLieuNopThue(masothue, macbql, sotiennop, ngaynop, kythue, tieumuc, muc);
+                                if ((System.IO.File.Exists(txtFilepath.Text)))
+                                {
+                                    System.IO.File.Delete(txtFilepath.Text);
+                                }
+                                if(lstErr.Count>0)
+                                    Export(lstErr);
+                                st.Append("$.notify('Thao tác thành công',{className: 'success',globalPosition: 'bottom right'});");
+                                ClientScript.RegisterClientScriptBlock(this.GetType(), "", st.ToString(), true);
+                            }
                         }
-                        st.Append("$.notify('Thao tác thành công',{className: 'success',globalPosition: 'bottom right'});");
-                        ClientScript.RegisterClientScriptBlock(this.GetType(), "", st.ToString(), true);
-
-
                     }
                     catch (Exception ms)
                     {
@@ -205,6 +361,10 @@ namespace TelerikWebApp1
             {
                 var data = ReadFromExcelfile(Session["File"].ToString());
                 grid.DataSource = data;
+            }
+            else
+            {
+                grid.DataSource = null;
             }
             if (grid.DataSource == null)
                 grid.DataSource = new string[] { };
